Add logger to bootstrapping. Rename old files
